+++ app/inputs/date_picker_input.rb
@@ -1,4 +1,4 @@
-class DatepickerInput < SimpleForm::Inputs::Base
+class DatePickerInput < SimpleForm::Inputs::Base
   def input
     template.content_tag(:div, class: 'input-group') do
       template.concat @builder.text_field(attribute_name, input_html_options)